Remove ControllerContentModelAliasDeprecation for deprecatingAlias
So, this is kind of a fun one.

ControllerContentModelAliasDeprecation was introduced, from what I
could tell, because in the 1.x days of Ember, Controllers had a
`content` property, which `model` was aliased to.
We decided to flip that, and make `model` the source of truth and
`content` the alias.
Problem though, until that flip, if someone overrode `content`,
`model` would break too, hence the addon.

Nowadays, it is now fine to override `content`, and we don't need
to warn anymore that it is deprecated. It is then, time to remove
it. Which I did. Removing code is the best.

We now bumped the deprecation up a notch and made it a deprecating
alias instead, and eventually we will remove the `content` alias.

Here's to removing yet more code in the future, and make Ember
sparkle-clean.